MAMP/phpMyAdmin test
Began testing using MAMP/phpmyAdmin. Set up a project, this folder needs to be saved in the MAMP- htdocs folder. Needs MAMPS installed as well as mySQL saved in the classpath. when going to http://localhost/csi3370g4/index.html you should be a basic interface based roughly on the design posed earlier